backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# AI Services
anthropic==0.42.0

# Local STT (no API key required — downloads ~466MB model on first run)
faster-whisper>=1.0.0

# Wake Word Detection (MIT, fully local, no API key)
openwakeword>=0.6.0

# TTS
edge-tts>=6.1.9

# Screen capture (replaces Electron desktopCapturer)
mss>=9.0.0

# Vision (Claude Haiku via anthropic SDK — no binary deps required)
Pillow==11.0.0

# Browser automation
playwright==1.49.0

# Notion integration
notion-client==2.2.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Configuration and Settings
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1
httpx[socks]
socksio

# Code Quality
black==24.10.0
isort==5.13.2
mypy==1.14.1
flake8==7.1.1

# Error Monitoring
sentry-sdk[fastapi]==2.19.2

# Utilities
python-json-logger==3.2.1
